Use -fno-strict-aliasing where needed
Gcc 6 became much more strict about strict aliasing: It now can catch
and warn about more illegal cases of type punning, and uses this assumption
more in optimizations, which can lead to incorrect code generation where
this assumption is violated.

In files which contain such violations, we need to use the
-fno-strict-aliasing compilation parameter, so gcc avoids making this
assumption in its optimizations (resulting in slightly less optimized code
for these files). The -Wno-strict-aliasing which we used previously is
not a good idea: it avoids warning about these cases, but continues to
generate bad code. We need to stop both, and this is what this patch does.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1469051748-11818-1-git-send-email-nyh@scylladb.com>