sched: tests for pin(thread*, cpu*)
Add tests for pin()ing this thread or a different thread.
I attempted to capture some of the edge cases, such as pin() on a running,
thread, a sleeping thread, or a thread which is often in migrate-disable
state, pinning to the cpu a thread is already on, and re-pinning an already
pinned thread.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1469628899-5537-5-git-send-email-nyh@scylladb.com>