test: re-add missing library for Boost-using tests
Some of our unit tests use the Boost filesystem library that in the past
we added to every image. But starting in commit 8693761 we no longer add
this library to every image, so we need to explicitly add it to the test
image. If we don't, tests like tst-rename.so (for example) fail with
missing boost::filesystem symbols.

Fixes #780
Refs #779

Signed-off-by: Nadav Har'El <nyh@scylladb.com>