net: Allow SIOCGIFBRDADDR even without IFF_BROADCAST
On Linux, one may query even a network interface that doesn't have the
"BROADCAST" flag with SIOCGIFBRDADDR - for example, the loopback
interface does not have this flag and yet in Linux, one may query it
for the broadcast address and it doesn't fail (the interfaces address is
returned).

Trying to query the broadcast address of an interface without one is not
very useful, but unfortunately musl's getifaddrs() implementation does
this (it assumes any non point-to-point interface has a broadcast address,
which is wrong for the loopback interface) and Linux is graceful about it,
and so should we, to be compatible with Linux.

With this patch (and the previous patch to another ioctl), the
getifaddrs() now works.

Fixes #783.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1470785000-22927-1-git-send-email-nyh@scylladb.com>