scripts/test.py: forgive old files in build directory
This patch fixes a failure report in our Jenkins build machine, which
includes a build directory that hasn't been "make clean"ed in 6 months:

test.py currently finds the list of tests to run by listing the files
in build/release/tests. But the test image Makefile (modules/tests/Makefile)
doesn't actually put everything in that directory into the test image -
rather it has a list of tests, and only those tests get listed in the
usr.manifest and therefore put into the image. If we decide to delete one
of the tests (as happened in commit 78854b4f0d32740f222869d4111461680f42ba3b)
and we do not "make clean", the old compiled test will remain in
build/release/tests - but not get included in the image, and test.py will
fail when it tries to run it.

Until commit 785e57e60969e25c19c87ad1b922ad9b66428c0e this problem was
benign, because test.py used to ignore missing object files. Because it
no longer does, we need to solve this problem.

The solution is for test.py to take the list of tests in the image from the
same usr.manifest which was used to build it.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>