osv::run(): run pthread TSD destructors before resetting app runtime
The change to osv::run() (and friends) to reuse the calling thread rather
than start a new one to run a new shared-object introduced a new problem:

The new shared object may use pthread_key_create() to add destructors to
run when closing threads. There is (currently) a global list of these
destructors and once added they never get removed (even if the object is
unloaded). The danger is that osv::run() runs code which adds such a
destructors and allocates data (with pthread_setspecific()) and then the
program is unloaded - but the thread will only exit (and the destructors
run) later - and crash when trying to run the destructor pointing to
already-deleted code.

Previously we didn't have this problem, because osv::run() had a new thread,
which exited (and ran the destructors) *before* the reference counter on
the application dropped to 0 and the shared object was unloaded.

A hacky workaround added in this patch is to run in
application::start_and_join - used by osv::run, application::run_and_join
and osv_execve - a new function which runs all these destructors. This
assumes the thread starting these run functions did not have any thread-
specific data (it would be deleted too!). It's a hack, but should be good
enough for use cases like osv_execve() which anyway uses a fresh thread
for running the application.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1472387362-11913-1-git-send-email-nyh@scylladb.com>