Partial fix for syscall instruction
We need to set up the LSTAR MSR and friends on each processor - it is not
enough to do it on processor 0. Otherwise, a system call instruction will
always fails if happens to run on any vcpu other than the first one.

With this patch, tst-syscall now passes on my machine with the release
compilation, on both single vcpu (run.py -c1) and many (-c4).
With debug build, the patch still fails - I have other fixes to make it
work.

Refs #791.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1472464153-3523-1-git-send-email-nyh@scylladb.com>
Signed-off-by: Pekka Enberg <penberg@scylladb.com>