syscall: save also %rbp and align stack
Not saving and restoring the rbp register causes tst-syscall to crash in
the debug build. With this patch, the debug build of this test no longer
crashes.

Once we do save %rbp, let's kill two birds in one stone, and also enable
backtrace_safe() (e.g., on abort) to go through the syscall_entry function
correctly. To do this, we need to set up the old-style frame pointer -
which means we need to push to the stack the return address (which we get
in %rcx), then the old %rbp, and then set %rbp to our %rsp.

Now there's an extra complication: Adding an odd number of 8-byte items
to the stack makes it, in my debug-build of the test, no longer 16-bytes
aligned. According to the C ABI, the stack must be 16-byte aligned when
calling a C function (syscall_wrapper()) - and the debug build has some
FPU-saving code which makes this assumptions, and crashes with #GP if not.

So we add in this patch also code to align the stack to 16 bytes before
calling the C function. We use a nice trick to do that without using up
another register.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1472474536-31851-1-git-send-email-nyh@scylladb.com>