syscall: skip red zone
The AMD64 ABI allows a function to use 128 bytes beyond the bottom of the
stack without updating the stack pointer - as an optimization known as the
"red zone". If this function calls another function, it is supposed to
update the stack pointer first; But with the SYSCALL instruction, it is not
aware it ends up calling a function.

So our implementation of syscall needs to skip 128 bytes of the stack,
which may be used by the caller's red zone, before making any use of the stack.

I haven't been able to reproduce a failure without this patch, but I do
believe it is necessary.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1472480089-10155-1-git-send-email-nyh@scylladb.com>