syscall: save/restore FPU state on syscall
As issue #798 explains, if an application uses floating point and then,
in the same function, also calls the SYSCALL instruction, the compiler
will not realize it needs to save the caller-saved floating-point state
(because it thinks it calls an x86 instruction, not a function).
So we need to do this saving in the SYSCALL implementation, before we
go on calling OSv functions which might clobber the FPU state.

The patch Uses fpu_lock to do the work of saving and restoring FPU state.

The case where this fix is actually needed is rare (in fact, we never
saw it in practice): Most applications do not in fact mix FPU use and
SYSCALL calling in a single function. But nevertheless, we add it to be
safe in case some application ever does it.

Unfortunately, the extra FPU saving means we slow down every SYSCALL
instruction. Note that it only slows down SYSCALL instructions, which
most applications do not use (golang is the notable exception), and
does NOT slow down the normal OSv ABI (the call to "system calls" as
ordinary functions). Also, we do not slow down the vDSO ABI which
golang uses instead of SYSCALL for system calls which are supposed to
be really quick.

Signed-off-by: Benoît Canet <benoit@scylladb.com>
Message-Id: <1477570022-14541-11-git-send-email-benoit@scylladb.com>
Signed-off-by: Nadav Har'El <nyh@scylladb.com>