virtio-rng: don't hang while waiting for entropy
random_kthread() has a loop, where it tries to get some entropy from
all sources (including the ring of internal events), feeds it into the
entropy pool, and then sleeps a bit and tries again, ad infinitum.

If the read function of *one* of the entropy sources blocks, the entire
random_kthread() blocks, and NONE of the entropy sources get used - not
even the internal one. This is bad. For example, if one of the entropy
sources blocks during boot, none of the plenty of internal random events
that happen during boot get used, and when the user's application uses
/dev/urandom on startup, it gets blocked because the entropy pool has not
been filled.

So the read function of live entropy sources should never block. In fact
a comment in live_entropy_source_feed() suggests that "This should be
quick, since it's a live entropy source". An entropy source which cannot
at this point provide any entropy should return zero bytes, instead of
blocking until it has some. The caller will find random bytes from other
sources, and doesn't want to wait for this specific source!

So virtio-rng's read function shouldn't wait for random bytes, it should
just provide whatever bytes it has right now.

Fixes #805

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <1479371945-17239-1-git-send-email-nyh@scylladb.com>
Reviewed-by: Pekka Enberg <penberg@scylladb.com>