printf(): fix "L" and "ll" prefixes
Musl, like Posix, allows the "L" prefix to be used only on long doubles,
and "ll" prefix on long long int. However, glibc actually allows both on
both types.

So this patch changes both prefixes to allow both types, to work like Linux.
It also adds a test that this works.

We saw this problem in an actual application, which used the "%Lx" format,
that worked on Linux but not on OSv (on OSv, the entire printf did not
happen because of the format parse error).

Fixes #818.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20161214163437.14901-2-nyh@scylladb.com>