build: possibly use cross-compilation strip
In the previous patch, scripts/mkbootfs.py started to use 'strip'.
However, this is run from the Makefile which allows cross-compilation,
and in that case, we need to use $STRIP instead of 'strip'.

So in this patch, we export the correct $STRIP from the Makefile to
its children, and use it in scripts/mkbootfs.py.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>