makefile: remove space from cross-compilation prefix
The spaces in $(if ...) are not ignored - they result in a space
in the beginning of the CROSS_PREFIX variable. This is mostly harmless,
except that mkbootfs.py now uses $STRIP and because it passes it to
popen, not the shell, it doesn't like that it begins with a space.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>