dhcp: fix wake of long-irrelevant thread.
Our DHCP API sets the "_waiter" variable to point to the current thread
when waiting for a DHCP response. However, when we finish waiting, we need
to clear this variable, otherwise we can find ourselves a long time later
(e.g., when a spurious DCHP packet arrives later) waking this thread,
which may no longer exist.

This bug caused actual random crashes, as described in issue #835.

Fixes #835

Signed-off-by: Nadav Har'El <nyh@scylladb.com>