app: clear up the osv::run() confusion
Before this patch we had two unrelated variants of osv::run() - one defined
in core/run.cc which runs a program in the current thread, and one defined
in app.cc which runs it in the background.

I have many reservations on how the second function works - especially
because of commit e270cf9c which prevented cleanup of completed
applications. But what bothers me even more is that these two unrelated
functions are confusingly have exactly the same name, with only slightly
different parameter types to give any hint which one is called.

So this patch changes the name of the second function to osv::run_background().

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170126131108.29596-1-nyh@scylladb.com>