sched::with_thread_by_id()
The existing function sched::thread::find_by_id() efficiently finds a
sched::thread object given a numerical id. However, if this call races
with the thread's exit, the returned thread pointer may no longer be
valid and using it might result in a crash.

This patch adds a safer function, sched::with_thread_by_id() which runs
a given function on the thread with the given id - while holding a mutex
which prevents threads (including this one) from exiting. Because it holds
this mutex, the function should typically be very short - e.g., test
something about this thread.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170201230213.23072-1-nyh@scylladb.com>