makefile: avoid known warning for lgamma*
In musl/src/math/lgamma*.c we have a known warning wrongly complaining
about a "maybe uninitialized" variable.

We already made sure this warning doesn't cause a compilation error, but
let's not show this warning at all.

Instead, we could have fixed the source file to not cause this warning.
But we took this file unmodified from Musl, and don't want to modify it
just to get rid of a warning.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170202145058.26076-1-nyh@scylladb.com>