gdb: search for binary with help of usr.manifest
When gdb searches for binary in host filesystem, only binary basename
is searched for. If multiple applications use the same name (say we have
OSV_DIR/app1/hello.so and OSV_DIR/app2/hello.so), then simply the first
found file is used. That might not be correct.

In cases when usr.manifest is available, it can be used to find binary
more reliably. This is implemented by that commit. If usr.manifest isn't
found or doesn't contain info about our binary, the old method of
traversing selected directories is used.

Fixes #848

Signed-off-by: Justin Cinkelj <justin.cinkelj@xlab.si>
Message-Id: <20170202214420.2729-1-justin.cinkelj@xlab.si>