line-discipline: move and update comment
The code in drivers/line-discipline.cc moved around a lot, and as a result
the comment trying to explain the purpose of this file got moved to the
middle of the file, and referred to no longer existing functions.

This patch moves and updates the comment. No actual code changes.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170313130650.20389-1-nyh@scylladb.com>