SIOCGIFFLAGS ioctl: remove BSD-specific bits
The SIOCGIFFLAGS ioctl returns a set of flags for a network interface.
Many of the classic bits there, like IFF_UP and IFF_LOOPBACK, are the same
in our internal implementation (BSD-based) and desired output (Linux),
but not all of them are the same. We already had a function to convert
the internal bitmask into a Linux bitmask, but it was incomplete.

Of particular importance is BSD's IFF_SIMPLEX flag, which does not exist
on Linux, and is set by OSv's Xen network driver. As we failed to clear
this flag, clients thought they were seeing Linux's IFF_SLAVE (which is
at the same bit position, 0x800). In particular OpenMPI did not like
seeing this flag, and refused to use this network interface.

Fixes #862

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170314133315.18971-1-nyh@scylladb.com>